fixed linter bugs

+++ app/Http/Controllers/TaskController.php
@@ -146,7 +146,9 @@ class TaskController extends Controller
 
         $toManyLabelsTask = LabelTask::select('label_id')->where('task_id', $task['id'])->get()->toArray();
 
-        if (!empty($labels = $request->get('labels'))) {
+        $labels = $request->get('labels');
+
+        if (count($labels) > 0) {
             array_map(fn($label) => !in_array($label, $toManyLabelsTask, true) ? LabelTask::insert([
                 'label_id' => $label,
                 'task_id' => $task->id]) : true, $labels);
@@ -167,7 +169,7 @@ class TaskController extends Controller
             return abort(403);
         }
         $task = Task::find($id);
-        if (Auth::id() === $task['created_by_id']) {
+        if (array_key_exists('created_by_id', $task) && Auth::id() === $task['created_by_id']) {
             $task->delete();
             flash('Задача успешно удалена')->success();
             return redirect('/tasks');
